move fs container to a new file and rename it
