+++ src/lib/core/overlay/overlay-container.ts
@@ -30,46 +30,3 @@ export class OverlayContainer {
     this._containerElement = container;
   }
 }
-
-/**
- * The OverlayContainer is the container in which all overlays will load.
- * It should be provided in the root component to ensure it is properly shared.
- */
-@Injectable()
-export class FullscreenFriendlyOverlayContainer extends OverlayContainer {
-  protected _createContainer(): void {
-    super._createContainer();
-
-    this._addFullscreenChangeListener(() => this._adjustParentForFullscreenChange());
-  }
-
-  private _adjustParentForFullscreenChange(): void {
-    if (!this._containerElement) {
-      return;
-    }
-
-    let fullscreenElement = this._getFullscreenElement();
-    let parent = fullscreenElement || document.body;
-    parent.appendChild(this._containerElement);
-  }
-
-  private _addFullscreenChangeListener(fn: () => void) {
-    if (document.fullscreenEnabled) {
-      document.addEventListener('fullscreenchange', fn);
-    } else if (document.webkitFullscreenEnabled) {
-      document.addEventListener('webkitfullscreenchange', fn);
-    } else if ((document as any).mozFullScreenEnabled) {
-      document.addEventListener('mozfullscreenchange', fn);
-    } else if ((document as any).msFullscreenEnabled) {
-      document.addEventListener('MSFullscreenChange', fn);
-    }
-  }
-
-  private _getFullscreenElement(): Element {
-    return document.fullscreenElement ||
-        document.webkitFullscreenElement ||
-        (document as any).mozFullScreenElement ||
-        (document as any).msFullscreenElement ||
-        null;
-  }
-}
